небольшие изменения в репозиториях

+++ src/main/java/com/sport/project/dao/repository/DisciplineRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DisciplineRepository extends JpaRepository<DisciplineEntity, Long> {
+public interface DisciplineRepository extends JpaRepository<DisciplineEntity, Integer> {
 
     /*Не добавила поиск по айди, потому что вроде как все основные методы в
     абстрактном интерфейсе
