+++ src/main/java/com/sport/project/dao/repository/LessonsRepository.java
@@ -4,6 +4,6 @@ import com.sport.project.dao.entity.LessonsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 //В документации нет, но может еще понадобится
-public interface LessonsRepository extends JpaRepository<LessonsEntity, Long> {
+public interface LessonsRepository extends JpaRepository<LessonsEntity, Integer> {
 
 }
